Remove old search classes
